+++ Cargo.toml
@@ -112,6 +112,9 @@ thiserror = "2.0"
 tls_codec = "0.4.1"
 tokio = { version = "1.43.0", default-features = false }
 tokio-stream = { version = "0.1", default-features = false }
+tokio-util = { version = "0.7", default-features = false, features = [
+  "compat",
+] }
 uuid = "1.12"
 vergen-git2 = "1.0.2"
 web-time = "1.1"
