+++ xmtp_mls/src/groups/device_sync/archive.rs
@@ -1,9 +1,55 @@
+pub use importer::ArchiveImporter;
+use thiserror::Error;
+use xmtp_common::time::now_ns;
+use xmtp_proto::xmtp::device_sync::{BackupElementSelection, BackupMetadataSave, BackupOptions};
+
 // Increment on breaking changes
 const BACKUP_VERSION: u16 = 0;
 
-mod export;
 mod export_stream;
-mod import;
+pub mod exporter;
+pub mod importer;
+
+#[derive(Debug, Error)]
+pub enum ArchiveError {
+    #[error("Missing metadata")]
+    MissingMetadata,
+}
+
+#[derive(Default)]
+pub struct BackupMetadata {
+    pub backup_version: u16,
+    pub elements: Vec<BackupElementSelection>,
+    pub exported_at_ns: i64,
+    pub start_ns: Option<i64>,
+    pub end_ns: Option<i64>,
+}
+
+impl BackupMetadata {
+    fn from_metadata_save(save: BackupMetadataSave, backup_version: u16) -> Self {
+        Self {
+            elements: save.elements().collect(),
+            end_ns: save.end_ns,
+            start_ns: save.start_ns,
+            exported_at_ns: save.exported_at_ns,
+            backup_version,
+        }
+    }
+}
+
+pub(crate) trait OptionsToSave {
+    fn from_options(options: BackupOptions) -> BackupMetadataSave;
+}
+impl OptionsToSave for BackupMetadataSave {
+    fn from_options(options: BackupOptions) -> BackupMetadataSave {
+        Self {
+            end_ns: options.end_ns,
+            start_ns: options.start_ns,
+            elements: options.elements,
+            exported_at_ns: now_ns(),
+        }
+    }
+}
 
 #[cfg(test)]
 mod tests {
@@ -14,7 +60,9 @@ mod tests {
         builder::ClientBuilder, groups::GroupMetadataOptions, utils::test::wait_for_min_intents,
     };
     use diesel::prelude::*;
+    use exporter::ArchiveExporter;
     use futures::io::Cursor;
+    use importer::ArchiveImporter;
     use std::{path::Path, sync::Arc};
     use xmtp_cryptography::utils::generate_local_wallet;
     use xmtp_db::group_message::MsgQueryArgs;
