--- xmtp_mls/src/groups/device_sync/archive/export_stream.rs
@@ -1,3 +1,0 @@
-pub(crate) mod consent_save;
-pub(crate) mod group_save;
-pub(crate) mod message_save;
